invoice view has been added , some changes have been added to clients and products views

+++ src/app/pages/products/product-list/product-list.component.ts
@@ -1,11 +1,13 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { GeneralService } from 'src/app/core/Services/general.service';
+import { ConfirmationService, MessageService } from 'primeng/api';
 
 @Component({
   selector: 'app-product-list',
   templateUrl: './product-list.component.html',
-  styleUrls: ['./product-list.component.scss']
+  styleUrls: ['./product-list.component.scss'],
+  providers: [ConfirmationService, MessageService]
 })
 export class ProductListComponent implements OnInit {
   dataAll: any[] = []
@@ -14,18 +16,20 @@ export class ProductListComponent implements OnInit {
   constructor(
     private generalService: GeneralService,
     public router: Router,
+    private confirmationService: ConfirmationService,
+    private messageService: MessageService
   ) { }
 
   ngOnInit() {
     this.getData()
   }
   getData() {
-    setTimeout(()=>{
+    setTimeout(() => {
       return this.generalService.getPagginated().subscribe((data) => {
-      this.dataAll = data
-    })
-    },100)
-    
+        this.dataAll = data
+      })
+    }, 100)
+
   }
   createNew() {
     this.router.navigate([`/main/${this.routerPath}/detail`]);
@@ -43,16 +47,26 @@ export class ProductListComponent implements OnInit {
       return this.getData()
     }
   }
-  requestDelete(id: string) {
-    return this.generalService.delete(id).subscribe((data) => {
-      this.getData()
+  requestDelete(event: Event, id: string) { 
+    this.confirmationService.close()
+    this.confirmationService.confirm({      
+      target: event.target as EventTarget,
+      message: 'Are you sure that you want to delete this?',
+      icon: 'pi pi-exclamation-triangle',
+      accept: () => {
+        return this.generalService.delete(id).subscribe((data) => {
+          this.getData()
+          this.messageService.add({ severity: 'info', summary: 'Confirmed', detail: 'The item has been deleted' })
+        })
+      },
+
     })
+    // return this.generalService.delete(id).subscribe((data) => {
+    //   this.getData()
+    // })
   }
   sumStocks(stocks: any[]) { 
-    let sum = 0
-    stocks.map(o=>sum+=o.quantity) 
-
-    return sum
+    return (stocks.reduce((x,y)=>x+y.quantity,0)) 
   }
 
 }
